adding CMU's Delphi data
Adding code for pulling data from Delphi API https://cmu-delphi.github.io/delphi-epidata/. Start date for Delphi data is set to a default date of March 7, 2020 to match other time-series data. Delphi is included in the default list of sources (JHU, Covid-Tracking) and takes 10 additional minutes to run given the total signals and dates we are pulling data for.